feat: refacto codes
